PreProcessTxt.cpp works
Wrote a C++ program that pre-processes the data input text file and
creates a shorter file containing only the lines of interest. The next
step will be to change the GetData class of CcsCal.py to call
PreProcessTxt then take in the pre-processed data file in directly and
not do any mass window filtering (this might be done in DtHist??).