began update to GetData object in CcsCal.py
added an input argument to the GetData initialization: pp=False. pp
starts as false and when its false it just does the conventional
method. When pp is set to a list of two values then it  calls the
PreProcessTxt.o program to generate a pre-processed input file then
uses numpy.genfromtxt to create an array with the mass, dtbin, and
intensity values… Now I need to add some options in the classes that
call GetData to actually use the new pp option