implemented pre-processing for every extraction in CcsCal.py
pre-processing is used for all of the drift time extractions in
CcsCal.py. It is MUCH faster but a bit jenky in terms the accuracy of
the drift times its pulling out. I think maybe something is going a bit
funny with the combining the dtbin data in the DtHist class (did I
delete those lines of code?) since in the plots i am seeing the
multiple curves and the fits look a little funky. I’ll just have to see…