updates to CcsCalibration class
began making updates to CcsCalibration class. So far i think I have
made most if not all of the necessary updates but I still have yet to
test any of those changes. Accordingly, the next step is to do a bunch
of testing to see if it works still and if (very likely, almost
inevitably) it does not then I can monkey around with things until it
does. ALSO maybe consider grouping some of the many arrays that there
are… determine which are initialized in groups and group them by that,
maybe not even necessarily by logical groupings. I’ll figure that out
at some point