added smoothing to GaussFit
implemented Savitsky-Golay filter to raw data in the GaussFit class to
improve the fit.
Right now the smooth window is set to 5 and the polynomial order is set
to 3 and the results look pretty good… I might play with them a bit
down the road to see if I can’t improve the accuracy.