updated GenerateReport class
the GenerateReport class is all up to date. Now all that is left to do
are some small TODO’s here and there and to benchmark the performance
and accuracy of this new approach. Qualitatively, using the
pre-processing is WAY faster than not using it.. I will have to see if
that holds up quantitatively.