more input updates
added another (a fourth) element to the input parameters list returned
by getInputParams. The aforementioned element is a numpy array from
numpy.genfromtxt containing the calibrant masses and lit ccs values as
well as the data file names and the masses for the compounds.

I think I will go ahead and create a class that unpacks all of this
data in a convenient to access manner